Refac a bit test/examples and refresh all curves at once in a graph
