+++ meson.build
@@ -14,7 +14,8 @@ extra_files=files(
     'Docker/Dockerfile',
     'Docker/build.sh',
     'README.md', 'COPYING',
-    'tests/sanity/signal_and_slots.sh'
+    'tests/sanity/signal_and_slots.sh',
+    'tests/manual-tests/common.py'
 )
 
 subdir('SciQLopPlots')
