+++ tests/manual-tests/common.py
@@ -1,0 +1,71 @@
+from PySide6.QtWidgets import QMainWindow, QApplication, QScrollArea,QWidget, QVBoxLayout, QTabWidget, QDockWidget
+from PySide6.QtCore import Qt
+import sys, os
+import numpy as np
+from datetime import datetime, timezone
+import threading
+from typing import Dict, Any
+
+from qtconsole.rich_jupyter_widget import RichJupyterWidget
+from qtconsole.inprocess import QtInProcessKernelManager
+
+os.environ['QT_API'] = 'PySide6'
+
+
+
+def fix_name(name):
+    return name.replace(" ", "_").replace(":", "_").replace("/", "_")
+
+class Tabs(QTabWidget):
+    def __init__(self,parent):
+        QTabWidget.__init__(self,parent)
+        self.setMouseTracking(True)
+        self.setTabPosition(QTabWidget.TabPosition.North)
+        self.setTabShape(QTabWidget.TabShape.Rounded)
+        self.setMovable(True)
+
+    def add_tab(self, widget, title):
+        self.addTab(widget, title)
+
+
+
+class MainWindow(QMainWindow):
+    def __init__(self):
+        QMainWindow.__init__(self)
+        self.setMouseTracking(True)
+        self.axisRects=[]
+        self.graphs = []
+        self._setup_kernel()
+        self._setup_ui()
+        self.setGeometry(400, 250, 542, 390);
+        self.setWindowTitle("SciQLopPlots Test")
+        self.add_to_kernel_namespace('main_window', self)
+
+    def add_tab(self, widget, title):
+        self.tabs.add_tab(widget, title)
+        self.add_to_kernel_namespace(fix_name(title), widget)
+
+    def _setup_kernel(self):
+        self.kernel_manager = QtInProcessKernelManager()
+        self.kernel_manager.start_kernel(show_banner=False)
+        self.kernel = self.kernel_manager.kernel
+        self.kernel.gui = 'qt'
+
+        self.kernel_client = self.kernel_manager.client()
+        self.kernel_client.start_channels()
+
+    def add_to_kernel_namespace(self, name, obj):
+        self.kernel.shell.push({name: obj})
+
+    def _setup_ui(self):
+        self.tabs = Tabs(self)
+        self.setCentralWidget(self.tabs)
+        self.ipython_widget = RichJupyterWidget(self)
+        dock = QDockWidget("IPython Console", self)
+        dock.setWidget(self.ipython_widget)
+        self.addDockWidget(Qt.BottomDockWidgetArea, dock)
+        self.ipython_widget.kernel_manager = self.kernel_manager
+        self.ipython_widget.kernel_client = self.kernel_client
+
+
+
